refactor(web): use axios to fetch issues
